+++ src/spatial_bvh.rs
@@ -66,7 +66,7 @@ pub struct SpatialBVH {
     guid: std::sync::OnceLock<String>,
     pub name: String,
     pub root: Option<Box<SpatialBVHNode>>,
-    pub world_size: f64,
+    pub world_size: f32,
     #[serde(skip)]
     pub object_guids: Vec<String>, // Parallel array to boxes - maps indices to GUIDs
     #[serde(skip)]
@@ -109,12 +109,12 @@ impl SpatialBVH {
     }
 
     /// Compute world size from bounding boxes
-    pub fn compute_world_size(bounding_boxes: &[OBB]) -> f64 {
+    pub fn compute_world_size(bounding_boxes: &[OBB]) -> f32 {
         if bounding_boxes.is_empty() {
             return 1000.0;
         }
 
-        let mut max_extent = 0.0f64;
+        let mut max_extent = 0.0f32;
         for bbox in bounding_boxes {
             // Find maximum absolute coordinate in any dimension
             let x_extent = (bbox.center[0] + bbox.half_size[0])
@@ -159,7 +159,7 @@ impl SpatialBVH {
         self.build(&bounding_boxes);
     }
 
-    pub fn from_boxes(bounding_boxes: &[OBB], world_size: f64) -> Self {
+    pub fn from_boxes(bounding_boxes: &[OBB], world_size: f32) -> Self {
         let mut bvh = Self::new();
         bvh.world_size = world_size;
         bvh.build(bounding_boxes);
@@ -583,7 +583,7 @@ impl SpatialBVH {
         &self,
         object_id: usize,
         bounding_boxes: &[OBB],
-        inflate: f64,
+        inflate: f32,
     ) -> Vec<usize> {
         if object_id >= bounding_boxes.len() {
             return Vec::new();
@@ -757,7 +757,7 @@ impl SpatialBVH {
         origin: &Point,
         direction: &Vector,
         aabb: &AABB,
-    ) -> Option<(f64, f64)> {
+    ) -> Option<(f32, f32)> {
         let min_x = aabb.cx - aabb.hx;
         let max_x = aabb.cx + aabb.hx;
         let min_y = aabb.cy - aabb.hy;
@@ -766,17 +766,17 @@ impl SpatialBVH {
         let max_z = aabb.cz + aabb.hz;
 
         let invx = if direction[0] == 0.0 {
-            f64::INFINITY
+            f32::INFINITY
         } else {
             1.0 / direction[0]
         };
         let invy = if direction[1] == 0.0 {
-            f64::INFINITY
+            f32::INFINITY
         } else {
             1.0 / direction[1]
         };
         let invz = if direction[2] == 0.0 {
-            f64::INFINITY
+            f32::INFINITY
         } else {
             1.0 / direction[2]
         };
@@ -870,7 +870,7 @@ pub fn expand_bits(v: u32) -> u32 {
     v
 }
 
-pub fn calculate_morton_code(x: f64, y: f64, z: f64, world_size: f64) -> u32 {
+pub fn calculate_morton_code(x: f32, y: f32, z: f32, world_size: f32) -> u32 {
     // Normalize coordinates to [0,1] range
     let nx = (x + world_size / 2.0) / world_size;
     let ny = (y + world_size / 2.0) / world_size;
